Extract methods to BenchmarkFileFormatsUtils
